Added Linting
Added Eslint
Code Cleanup
