+++ chrome/ebay/scrapDataEbay.js
@@ -89,7 +89,9 @@ export async function scrapDataEbay(activeListings, downloadImages) {
         });  
 
         if(activeListings){
-          chrome.runtime.sendMessage({ action: 'downloadEbayDesc', itemNumber: a.href.split('/')[4]});
+          //chrome.runtime.sendMessage({ action: 'downloadEbayDesc', itemNumber: a.href.split('/')[4]});
+          chrome.runtime.sendMessage({ action: 'queueShippingInfo', itemNumber: a.href.split('/')[4]});
+
           if(downloadImages){
             chrome.runtime.sendMessage({ action: 'downloadEbayImage', itemNumber: a.href.split('/')[4]});           
           }  
@@ -186,7 +188,7 @@ export async function scrapEbayImages(itemNumber, downloadImages) {
   function checkReadyState() {
     return new Promise((resolve, reject) => {
       if(document.readyState === 'complete') {
-        retrieveImages().then(retrieveDescription).then(resolve);
+        retrieveImages().then(resolve);
       }else{
         setTimeout(() => checkReadyState().then(resolve), 1000);
       }
@@ -223,63 +225,6 @@ export async function scrapEbayImages(itemNumber, downloadImages) {
    await checkReadyState();
 }
 
-export async function scrapEbayPostage(itemNumber) {
-  function checkReadyState() {
-    return new Promise((resolve, reject) => {
-      let timeoutId = setTimeout(() => {
-        clearTimeout(timeoutId);
-        reject(new Error('Page load timed out after 10 seconds'));
-      }, 10000); // 10 seconds timeout
-  
-      function check() {
-        if(document.readyState === 'complete') {
-          clearTimeout(timeoutId);
-          console.log('readyState is complete');
-          retirevePostage(); 
-          resolve();
-        } else {
-          console.log('readyState is not complete');
-          setTimeout(check, 1000);
-        }
-      }
-  
-      check();
-    });
-  }
-
-  function retirevePostage() {
-    return new Promise((resolve, reject) => {
-      function checkPostageElement() {
-          let majorElement = document.querySelector('input[name="majorWeight"]');
-          let minorElement = document.querySelector('input[name="minorWeight"]');
-          let packageLength = document.querySelector('input[name="packageLength"]');
-          let packageWidth = document.querySelector('input[name="packageWidth"]');
-          let packageHeight = document.querySelector('input[name="packageHeight"]');
-
-          if (majorElement) {
-            console.log('retirevePostage');
-            console.log(majorElement.innerText);
-            chrome.runtime.sendMessage({ 
-              action: 'updatePostage', 
-              majorElement: majorElement.value, 
-              minorElement: minorElement.value,
-              packageLength: packageLength.value,
-              packageWidth: packageWidth.value,
-              packageHeight: packageHeight.value,
-              item: itemNumber});
-            resolve();
-        } else {
-          setTimeout(checkPostageElement, 1000); // wait for 1 second before checking again
-        }
-      }
-      checkPostageElement();
-    }).catch((e) => {
-      console.log(e);
-      reject();
-    });
-  }
-}
-
 export async function scrapEbayDescriptions(itemNumber) {
   
   function checkReadyState() {
@@ -308,21 +253,23 @@ export async function scrapEbayDescriptions(itemNumber) {
 
   function retrieveDescription() {
     return new Promise((resolve, reject) => {
-      function checkDescElement() {
-          let descElement = document.querySelector('div[data-testid="x-item-description-child"]');
-          console.log('retrieveDescription');
-          console.log(descElement); 
-          if (descElement) {
-            chrome.runtime.sendMessage({ action: 'updateDesc', desc: descElement.innerText, item: itemNumber});
-            resolve();
-        } else {
-          setTimeout(checkDescElement, 1000); // wait for 1 second before checking again
+      try{
+        function checkDescElement() {
+            let descElement = document.querySelector('div[data-testid="x-item-description-child"]');
+            console.log('retrieveDescription');
+            console.log(descElement); 
+            if (descElement) {
+              chrome.runtime.sendMessage({ action: 'updateDesc', desc: descElement.innerText, item: itemNumber});
+              resolve();
+          } else {
+            setTimeout(checkDescElement, 1000); // wait for 1 second before checking again
+          }
         }
+        checkDescElement();
+      }catch(e){
+        console.log(e);
+        reject();   
       }
-      checkDescElement();
-    }).catch((e) => {
-      console.log(e);
-      reject();
     });
   }
