+++ chrome/eslint.config.mjs
@@ -1,0 +1,19 @@
+import globals from "globals";
+import pluginJs from "@eslint/js";
+
+
+export default [
+    {
+        languageOptions: { 
+        globals: {    
+            ...globals.browser, 
+            "chrome": "readonly",
+        },
+        parserOptions: {
+            // Required for certain syntax usages
+            "ecmaVersion": 2020
+            }, 
+        }
+    },
+    pluginJs.configs.recommended,
+];
