+++ chrome/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "chrome",
+  "version": "1.0.0",
+  "description": "Sales Auction Extentsion",
+  "main": "background.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@eslint/js": "^9.3.0",
+    "eslint": "^9.3.0",
+    "globals": "^15.3.0"
+  }
+}
